move files to logic folder
